Gary's Boxen Setup
This is a giant commit comprised of all my Boxen config settings.  I had
multiple commits, but it was a giant bitch to keep the repo up-to-date with
the main Boxen repo.  I don't care about history.  Ship it.

Refactor subclasses and variables

I want to keep all variables in the params class and
keep everything neat and tiday.

Refactor: Use an array for homebrew packages

Ignore /vendor/puppet/cache...

...so git doesn't try to add all vendored modules to the repo.